+++ x/uibc/quota/keeper/keeper_test.go
@@ -4,6 +4,10 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/cosmos/cosmos-sdk/codec"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
+	porttypes "github.com/cosmos/ibc-go/v6/modules/core/05-port/types"
+
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/simapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -17,6 +21,7 @@ import (
 
 	umeeapp "github.com/umee-network/umee/v4/app"
 	appparams "github.com/umee-network/umee/v4/app/params"
+	"github.com/umee-network/umee/v4/tests/tsdk"
 	"github.com/umee-network/umee/v4/x/uibc"
 	"github.com/umee-network/umee/v4/x/uibc/quota/keeper"
 )
@@ -85,4 +90,18 @@ func initKeeperTestSuite(t *testing.T) *KeeperTestSuite {
 	s.msgServer = keeper.NewMsgServerImpl(app.UIbcQuotaKeeper)
 
 	return s
+}
+
+// creates keeper with all external dependencies (app, leverage etc...)
+func initFullKeeper(
+	t *testing.T,
+	cdc codec.BinaryCodec,
+	ics4Wrapper porttypes.ICS4Wrapper,
+	leverageKeeper uibc.LeverageKeeper,
+	oracleKeeper uibc.Oracle,
+) (sdk.Context, keeper.Keeper) {
+	storeKey := storetypes.NewMemoryStoreKey("quota")
+	k := keeper.NewKeeper(cdc, storeKey, ics4Wrapper, leverageKeeper, oracleKeeper)
+	ctx, _ := tsdk.NewCtxOneStore(t, storeKey)
+	return ctx, k
 }